+++ EurexEti/Messages/RetransmitResponse.cs
@@ -54,7 +54,7 @@ namespace Eurex.EtiDerivatives.v121
 
             // --- complete header ---
 
-            BodyLen.Encode(pointer, offset, current - offset);
+            BodyLen.Encode(pointer, offset, (uint)(current - offset));
         }
 
         /// <summary>
