+++ EurexEti/Messages/SrqsInquireSmartRespondentRequest.cs
@@ -62,7 +62,7 @@ namespace Eurex.EtiDerivatives.v121
 
             // --- complete header ---
 
-            BodyLen.Encode(pointer, offset, current - offset);
+            BodyLen.Encode(pointer, offset, (uint)(current - offset));
         }
 
         /// <summary>
